Move `ParsedLogRecord`
